add change cart

+++ src/main/java/dto/Cart.java
@@ -9,12 +9,12 @@ public class Cart {
     @Column(name="id")
     @GeneratedValue(strategy= GenerationType.IDENTITY)
 
-    private Integer id;
+    private Long id;
     int order_id;
     int product_count;
     int price_id;
 
-    public Cart(Integer id,
+    public Cart(Long id,
                 Integer order_id,//ValueFromDB,
                 Integer product_count, //ValueFromDB,
                 Integer price_id //ValueFromDB
@@ -29,11 +29,11 @@ public class Cart {
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
